Reclaim GrooveBot name.
